+++ Code/Utils/JLUtils.cs
@@ -10,36 +10,77 @@ using TinyJson;
 namespace JLPlugin.Utils
 {
     using Data;
+    using System;
 
     public static class JLUtils
     {
         public static void LoadCardsFromFiles()
         {
-            foreach ( string file in Directory.EnumerateFiles( Paths.PluginPath, "*.jldr", SearchOption.AllDirectories ) )
+            foreach ( string file in Directory.EnumerateFiles( Paths.PluginPath, "*.jldr", SearchOption.AllDirectories ).Where( file => !file.EndsWith( "_encounter.jldr" ) ) )
             {
                 string filename = file.Substring( file.LastIndexOf( Path.DirectorySeparatorChar ) + 1 );
 
-                Data.CardData card = CreateFromJSON( File.ReadAllText( file ) );
+                string text = File.ReadAllText( file );
 
-                if ( card is null )
+                if ( !text.ToLowerInvariant().StartsWith( "#ignore" ) && !text.ToLowerInvariant().StartsWith( "# ignore" ) )
                 {
-                    Plugin.Log.LogWarning( $"Failed to load { filename }" );
+                    Data.CardData card = CreateCardFromJSON( text );
 
-                    continue;
+                    if ( card is null )
+                    {
+                        Plugin.Log.LogWarning( $"Failed to load card { filename }" );
+
+                        continue;
+                    }
+                    else
+                    {
+                        if ( !filename.EndsWith( "_card.jldr" ) )
+                        {
+                            Plugin.Log.LogWarning( $"Card { filename } does not have the '_card' postfix. Cards without this postfix will be unsupported in the future." );
+                        }
+
+                    }
+
+                    if ( card.fieldsToEdit is null )
+                    {
+                        card.GenerateNew();
+                        continue;
+                    }
+
+                    card.Edit();
                 }
-
-                if ( card.fieldsToEdit is null )
-                {
-                    card.GenerateNew();
-                    continue;
-                }
-
-                card.Edit();
             }
         }
 
-        public static CardData CreateFromJSON( string jsonString )
+        public static void LoadEncountersFromFiles()
+        {
+            foreach ( string file in Directory.EnumerateFiles( Paths.PluginPath, "*_encounter.jldr", SearchOption.AllDirectories ) )
+            {
+                string filename = file.Substring( file.LastIndexOf( Path.DirectorySeparatorChar ) + 1 );
+
+                string text = File.ReadAllText( file );
+
+                if ( !text.ToLowerInvariant().StartsWith( "#ignore" ) && !text.ToLowerInvariant().StartsWith( "# ignore" ) )
+                {
+                    Data.CustomEncounterData encounter = CreateEncounterFromJSON(  text );
+
+                    if ( encounter is null )
+                    {
+                        Plugin.Log.LogWarning( $"Failed to load encounter { filename }" );
+
+                        continue;
+                    }
+
+                    encounter.GenerateNew();
+                }
+            }
+        }
+
+        public static CardData CreateCardFromJSON( string jsonString )
             => JSONParser.FromJson<CardData>( jsonString );
+
+        public static CustomEncounterData CreateEncounterFromJSON( string jsonString )
+            => JSONParser.FromJson<CustomEncounterData>( jsonString );
 
         public static Texture2D LoadTexture2D( string image )
         {